web-sender/src/main/java/org/openmrs/eip/web/sender/SenderSyncMessageController.java
package org.openmrs.eip.web.sender;

import java.util.Map;

import org.openmrs.eip.app.management.entity.SenderSyncMessage;
import org.openmrs.eip.web.RestConstants;
import org.openmrs.eip.web.contoller.BaseRestController;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping(RestConstants.API_PATH + "/dbsync/sender/sync")
public class SenderSyncMessageController extends BaseRestController {
	
	private static final Logger log = LoggerFactory.getLogger(SenderSyncMessageController.class);
	
	@Override
	public Class<?> getClazz() {
		return SenderSyncMessage.class;
	}
	
	@GetMapping
	public Map<String, Object> getAll() {
		if (log.isDebugEnabled()) {
			log.debug("Fetching sender sync messages");
		}
		
		return doGetAll();
	}
	
	@GetMapping("/{id}")
	public Object get(@PathVariable("id") Long id) {
		if (log.isDebugEnabled()) {
			log.debug("Fetching sender sync message with id: " + id);
		}
		
		return doGet(id);
	}
	
}
